+++ backend/perc3d.c
@@ -53,7 +53,7 @@ float Res = DEFAULTRESOLUTION;
 /* VCCTL software version used to create input file */
 float Version;
 
-long int Ntot = 0, Nperc = 0, Nburnt = 0;
+int Ntot = 0, Nperc = 0, Nburnt = 0;
 FILE *Resfile;
 
 int burn3d(int npix, int d1, int d2, int d3);
@@ -96,7 +96,7 @@ int main(void) {
    *	Allocate memory for Mic array
    ***/
 
-  Mic = sibox((long)Xsyssize, (long)Ysyssize, (long)Zsyssize);
+  Mic = sibox(Xsyssize, Ysyssize, Zsyssize);
   if (!Mic) {
     fclose(infile);
     bailout("perc3d", "Could not allocate memory for Mic array");
@@ -171,7 +171,7 @@ int burn3d(int npix, int d1, int d2, int d3) {
   int *nmatx, *nmaty, *nmatz, *nnewx, *nnewy, *nnewz;
   int mult1, mult2, npix1, npix2, npix3;
   int xl, xh, px, py, pz, qx, qy, qz, xcn, ycn, zcn;
-  long int ntop, nthrough, ncur, nnew, nphc;
+  int ntop, nthrough, ncur, nnew, nphc;
 
   npix1 = npix;
   npix2 = npix;
@@ -548,8 +548,8 @@ int burn3d(int npix, int d1, int d2, int d3) {
   }
 
   printf("Phase ID = %d \n", npix);
-  printf("Number accessible from first surface = %ld \n", ntop);
-  printf("Number contained in through pathways= %ld \n", nthrough);
+  printf("Number accessible from first surface = %d \n", ntop);
+  printf("Number contained in through pathways= %d \n", nthrough);
   fflush(stdout);
 
   Ntot = nphc;
@@ -558,7 +558,7 @@ int burn3d(int npix, int d1, int d2, int d3) {
   if (nthrough > 0)
     bflag = 1;
 
-  fprintf(Resfile, "%ld %ld %ld \n", Ntot, Nperc, Nburnt);
+  fprintf(Resfile, "%d %d %d \n", Ntot, Nperc, Nburnt);
 
   free(nmatx);
   free(nmaty);
